+++ CHANGELOG.md
@@ -10,12 +10,15 @@
   - com.dwarfeng.judge.stack.bean.entity.AdapterInfo。
   - com.dwarfeng.judge.stack.bean.entity.AdapterSupport。
 
+- 实现预设过滤器。
+  - com.dwarfeng.judge.impl.handler.filter.groovy.GroovyFilterRegistry。
 
 - 实现预设适配器。
   - com.dwarfeng.judge.impl.handler.adapter.groovy.GroovyAdapterRegistry。
 
 - 实现核心机制。
   - 适配机制。
+  - 过滤机制。
 
 - 重构核心机制。
   - 提供机制。